+++ sdk/messaging/azservicebus/internal/namespace.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
+	"net"
 	"runtime"
 	"sync"
 	"time"
@@ -20,7 +21,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/internal/rpc"
 	"github.com/Azure/go-amqp"
 	"github.com/devigned/tab"
-	"nhooyr.io/websocket"
 )
 
 const (
@@ -35,7 +35,8 @@ type (
 		TokenProvider *sbauth.TokenProvider
 		tlsConfig     *tls.Config
 		userAgent     string
-		useWebSocket  bool
+
+		newWebSocketConn func(ctx context.Context, args NewWebSocketConnArgs) (net.Conn, error)
 
 		baseRetrier Retrier
 
@@ -107,10 +108,19 @@ func NamespaceWithUserAgent(userAgent string) NamespaceOption {
 	}
 }
 
+// NewWebSocketConnArgs are the arguments to the NewWebSocketConn function you pass if you want
+// to enable websockets.
+type NewWebSocketConnArgs struct {
+	// NOTE: this struct is exported via client.go:NewWebSocketConnArgs
+
+	// Host is the the `wss://<host>` to connect to
+	Host string
+}
+
 // NamespaceWithWebSocket configures the namespace and all entities to use wss:// rather than amqps://
-func NamespaceWithWebSocket() NamespaceOption {
+func NamespaceWithWebSocket(newWebSocketConn func(ctx context.Context, args NewWebSocketConnArgs) (net.Conn, error)) NamespaceOption {
 	return func(ns *Namespace) error {
-		ns.useWebSocket = true
+		ns.newWebSocketConn = newWebSocketConn
 		return nil
 	}
 }
@@ -173,15 +183,14 @@ func (ns *Namespace) newClient(ctx context.Context) (*amqp.Client, error) {
 		)
 	}
 
-	if ns.useWebSocket {
-		wssHost := ns.getWSSHostURI() + "$servicebus/websocket"
-		opts := &websocket.DialOptions{Subprotocols: []string{"amqp"}}
-		wssConn, _, err := websocket.Dial(ctx, wssHost, opts)
+	if ns.newWebSocketConn != nil {
+		nConn, err := ns.newWebSocketConn(ctx, NewWebSocketConnArgs{
+			Host: ns.getWSSHostURI() + "$servicebus/websocket",
+		})
 
 		if err != nil {
 			return nil, err
 		}
-		nConn := websocket.NetConn(context.Background(), wssConn, websocket.MessageBinary)
 
 		return amqp.New(nConn, append(defaultConnOptions, amqp.ConnServerHostname(ns.FQDN))...)
 	}